requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dataclasses==0.6
drms==0.6.0
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
scikit-learn==0.22.2.post1
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
torch==1.7.0+cpu
torchvision==0.8.1+cpu
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
